pyproject.toml
[tool.poetry]
name = "sire"
version = "0.1.0"
description = "Proyecto Data Warehouse"
authors = ["Nelson Arias <nelcetreal@gmail.com>"]
readme = "README.md"
packages = [{include = "sire", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
python-dotenv = ">=1.0.0,<2.0.0"
apache-airflow = {extras = ["postgres", "oracle"], version = ">=2.8.1,<3.0.0"}
apache-airflow-providers-postgres = ">=5.6.0,<6.0.0"
psycopg = {extras = ["binary"], version = ">=3.1.10,<4.0.0"}
oracledb = "==2.0.0"
pyspark = "==3.5.5"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
cx_Oracle = "^8.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
